+++ lib/sidekiq/delay.rb
@@ -3,34 +3,38 @@
 require "sidekiq"
 
 module Sidekiq
-  module DelayExtensions
-    def self.enable_delay!
+  module Delay
+    class << self; attr_accessor :limit_payload_size; end
+
+    def self.enable_delay!(limit_payload_size: false)
+      self.limit_payload_size = limit_payload_size
       if defined?(::ActiveSupport)
-        require "sidekiq/delay_extensions/active_record"
-        require "sidekiq/delay_extensions/action_mailer"
+        require "sidekiq/delay/active_record"
+        require "sidekiq/delay/action_mailer"
 
         # Need to patch Psych so it can autoload classes whose names are serialized
         # in the delayed YAML.
-        Psych::Visitors::ToRuby.prepend(Sidekiq::DelayExtensions::PsychAutoload)
+        Psych::Visitors::ToRuby.prepend(Sidekiq::Delay::PsychAutoload)
 
         ActiveSupport.on_load(:active_record) do
-          include Sidekiq::DelayExtensions::ActiveRecord
+          include Sidekiq::Delay::ActiveRecord
         end
         ActiveSupport.on_load(:action_mailer) do
-          extend Sidekiq::DelayExtensions::ActionMailer
+          extend Sidekiq::Delay::ActionMailer
         end
       end
 
-      require "sidekiq/delay_extensions/class_methods"
-      Module.__send__(:include, Sidekiq::DelayExtensions::Klass)
+      require "sidekiq/delay/class_methods"
+      Module.include Sidekiq::Delay::Klass
 
-      require "sidekiq/delay_extensions/api"
-      Sidekiq::JobRecord.prepend(Sidekiq::DelayExtensions::JobRecord)
+      require "sidekiq/delay/api"
+      Sidekiq::JobRecord.prepend(Sidekiq::Delay::JobRecord)
     end
 
     module PsychAutoload
       def resolve_class(klass_name)
         return nil if !klass_name || klass_name.empty?
+
         # constantize
         names = klass_name.split("::")
         names.shift if names.empty? || names.first.empty?
@@ -43,4 +47,6 @@ module Sidekiq
       end
     end
   end
+
+  DelayExtensions = Delay
 end