+++ lib/sidekiq/delay_extensions/testing.rb
@@ -1,8 +1,0 @@
-# frozen_string_literal: true
-
-require "sidekiq/testing"
-
-module Sidekiq
-  Sidekiq::DelayExtensions::DelayedMailer.extend(TestingExtensions) if defined?(Sidekiq::DelayExtensions::DelayedMailer)
-  Sidekiq::DelayExtensions::DelayedModel.extend(TestingExtensions) if defined?(Sidekiq::DelayExtensions::DelayedModel)
-end