+++ myapp/app/mailers/user_mailer.rb
@@ -1,9 +1,11 @@
+# frozen_string_literal: true
+
 class UserMailer < ActionMailer::Base
-  default from: "sidekiq@example.com"
+  default from: 'sidekiq@example.com'
 
   def greetings(now)
     @now = now
     @hostname = `hostname`.strip
-    mail(:to => 'mperham@gmail.com', :subject => 'Ahoy Matey!')
+    mail(to: 'mperham@gmail.com', subject: 'Ahoy Matey!')
   end
 end