+++ myapp/app/models/exiter.rb
@@ -1,6 +1,8 @@
+# frozen_string_literal: true
+
 class Exiter
   def self.run
-    Sidekiq.logger.warn "Success"
+    Sidekiq.logger.warn 'Success'
     Thread.new do
       sleep 0.1
       exit(0)