+++ myapp/config/initializers/sidekiq.rb
@@ -1,30 +1,32 @@
+# frozen_string_literal: true
+
 Sidekiq.configure_client do |config|
-  config.redis = { :size => 2 }
+  config.redis = { size: 2 }
 end
+
 Sidekiq.configure_server do |config|
-  config.on(:startup) { }
-  config.on(:quiet) { }
+  config.on(:startup) {}
+  config.on(:quiet) {}
   config.on(:shutdown) do
-    #result = RubyProf.stop
+    # result = RubyProf.stop
 
     ## Write the results to a file
     ## Requires railsexpress patched MRI build
     # brew install qcachegrind
-    #File.open("callgrind.profile", "w") do |f|
-      #RubyProf::CallTreePrinter.new(result).print(f, :min_percent => 1)
-    #end
+    # File.open("callgrind.profile", "w") do |f|
+    # RubyProf::CallTreePrinter.new(result).print(f, :min_percent => 1)
+    # end
   end
 end
 
 class EmptyWorker
-  include Sidekiq::Worker
+  include Sidekiq::Job
 
-  def perform
-  end
+  def perform; end
 end
 
 class TimedWorker
-  include Sidekiq::Worker
+  include Sidekiq::Job
 
   def perform(start)
     now = Time.now.to_f
@@ -32,7 +34,7 @@ class TimedWorker
   end
 end
 
-Sidekiq::DelayExtensions.enable_delay!
+Sidekiq::Delay.enable_delay!
 
 module Myapp
   class Current < ActiveSupport::CurrentAttributes
@@ -40,5 +42,5 @@ module Myapp
   end
 end
 
-require "sidekiq/middleware/current_attributes"
+require 'sidekiq/middleware/current_attributes'
 Sidekiq::CurrentAttributes.persist(Myapp::Current) # Your AS::CurrentAttributes singleton