+++ test/extensions_test.rb
@@ -1,20 +1,13 @@
 # frozen_string_literal: true
+
 require_relative 'helper'
+require_relative 'models'
 require 'sidekiq/api'
-require 'active_record'
-require 'action_mailer'
-Sidekiq::DelayExtensions.enable_delay!
 
-describe Sidekiq::DelayExtensions do
-  before do
-    Sidekiq.redis {|c| c.flushdb }
-  end
+Sidekiq::Delay.enable_delay!
 
-  class MyModel < ActiveRecord::Base
-    def self.long_class_method
-      raise "Should not be called!"
-    end
-  end
+describe Sidekiq::Delay do
+  before { Sidekiq.redis(&:flushdb) }
 
   it 'allows delayed execution of ActiveRecord class methods' do
     assert_equal [], Sidekiq::Queue.all.map(&:name)
@@ -49,12 +42,6 @@ describe Sidekiq::DelayExtensions do
     assert_equal 1, ss.size
   end
 
-  class UserMailer < ActionMailer::Base
-    def greetings(a, b)
-      raise "Should not be called!"
-    end
-  end
-
   it 'allows delayed delivery of ActionMailer mails' do
     assert_equal [], Sidekiq::Queue.all.map(&:name)
     q = Sidekiq::Queue.new
@@ -78,11 +65,6 @@ describe Sidekiq::DelayExtensions do
     assert_equal 1, ss.size
   end
 
-  class SomeClass
-    def self.doit(arg)
-    end
-  end
-
   it 'allows delay of any ole class method' do
     q = Sidekiq::Queue.new
     assert_equal 0, q.size
@@ -90,12 +72,9 @@ describe Sidekiq::DelayExtensions do
     assert_equal 1, q.size
   end
 
-  module SomeModule
-    def self.doit(arg)
-    end
-  end
-
   it 'logs large payloads' do
+    Sidekiq::Delay.limit_payload_size = true
+
     output = capture_logging(Logger::WARN) do
       SomeClass.delay.doit('a' * 8192)
     end
@@ -108,5 +87,4 @@ describe Sidekiq::DelayExtensions do
     SomeModule.delay.doit(Date.today)
     assert_equal 1, q.size
   end
-
 end